+++ tests/conhead/test_main.py
@@ -1,7 +1,6 @@
 # Copyright 2022 Rafe Kaplan
 # SPDX-License-Identifier: Apache-2.0
 #
-# Updated: 2022-06-09
 import datetime
 import inspect
 import logging