+++ CHANGES.rst
@@ -2,11 +2,11 @@
     Copyright 2022 Rafe Kaplan
     SPDX-License-Identifier: Apache-2.0
 
-    Updated: 2022-06-06
 
 Next Version
 ============
 - Integration with readthe docs at https://conhead.readthedocs.io.
+- Add --show-changes which shows old header and new header.
 
 v0.4.0
 ======